refactor: [DX-1547] Rename Notification and Alert
