+++ optimus/lib/optimus.dart
@@ -44,11 +44,11 @@ export 'src/dropdown/embedded_search.dart';
 export 'src/enabled.dart';
 export 'src/expansion/expansion_tile.dart';
 export 'src/feedback/alert.dart';
+export 'src/feedback/alert_overlay.dart';
 export 'src/feedback/banner.dart';
 export 'src/feedback/feedback_link.dart';
 export 'src/feedback/feedback_variant.dart';
-export 'src/feedback/notification.dart';
-export 'src/feedback/notification_overlay.dart';
+export 'src/feedback/system_wide_banner.dart';
 export 'src/form/checkbox_form_field.dart';
 export 'src/form/date_input_field.dart';
 export 'src/form/date_input_form_field.dart';
